+++ src/main/webapp/resources/js/pages/dashboard.js
@@ -1,0 +1,8 @@
+import angular from "angular";
+import { EventsModule } from "../modules/events/events";
+import { DashboardAnnouncementsModule } from "./announcement/announcementDashboard";
+
+const app = angular.module("irida");
+app
+  .requires.push(EventsModule)
+  .requires.push(DashboardAnnouncementsModule);