+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changes
 * [UI]: Fixed URL for concatenation of sample sequence files. (0.18.1)
 * [UI]: Fix broken link for concatenating files. (0.18.2)
 * [UI]: Fix broken permissions for downloading reference files not associated with a project. (0.18.2)
+* [UI]: Fixed issued with layout of the events panel on the dashboard page. (0.18.2)
 
 0.17.0 to 0.18.0
 ----------------